cleanup reference project
Update tests
