+++ reference-project/src/main/java/com/devonfw/demoquarkus/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.devonfw.application.quarkus.sample.general;
+package com.devonfw.demoquarkus;
 
 import javax.ws.rs.core.Application;
 
@@ -9,9 +9,10 @@ import org.eclipse.microprofile.openapi.annotations.info.License;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 
 @OpenAPIDefinition(tags = {
-@Tag(name = "animal", description = "Animal operations.") }, info = @Info(title = "My Demo App RESTful API", version = "1.0.0", contact = @Contact(name = "API Support", email = "support@acme.com"), license = @License(name = "Apache 2.0", url = "https://www.apache.org/licenses/LICENSE-2.0.html")))
+@Tag(name = "animal", description = "Animal API.") }, info = @Info(title = "My Demo App RESTful API", version = "1.0.0", contact = @Contact(name = "API Support", email = "support@acme.com")))
 /**
  * JaxRS application is not required in quarkus, but it is useful to place central API docs somewhere.
+ * We could also use package-info.java for this.
  */
 public class DemoApplication extends Application {
 }
