+++ reference-project/src/main/java/com/devonfw/demoquarkus/rest/v1/mapper/AnimalMapper.java
@@ -1,18 +1,19 @@
-package com.devonfw.application.quarkus.sample.animalmanagement.common.api;
+package com.devonfw.demoquarkus.rest.v1.mapper;
 
 import java.util.stream.Stream;
 
+import com.devonfw.demoquarkus.domain.model.AnimalSearchCriteria;
+import com.devonfw.demoquarkus.domain.model.Animal;
+import com.devonfw.demoquarkus.rest.v1.model.AnimalSearchCriteriaDTO;
 import org.mapstruct.Mapper;
 import org.tkit.quarkus.jpa.daos.PageResult;
 import org.tkit.quarkus.rs.mappers.OffsetDateTimeMapper;
 import org.tkit.quarkus.rs.models.PageResultDTO;
 
-import com.devonfw.application.quarkus.sample.animalmanagement.dataaccess.api.Animal;
-import com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to.AnimalDTO;
-import com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to.AnimalSearchCriteriaDTO;
-import com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to.NewAnimalDTO;
+import com.devonfw.demoquarkus.rest.v1.model.AnimalDTO;
+import com.devonfw.demoquarkus.rest.v1.model.NewAnimalDTO;
 
-//mapstruct will generate an impl class from this interface at compile time
+//mapstruct will generate an impl class(CDI bean, see pom.xml) from this interface at compile time
 @Mapper(uses = OffsetDateTimeMapper.class)
 public interface AnimalMapper {
