+++ reference-project/src/main/java/com/devonfw/demoquarkus/rest/v1/model/AnimalDTO.java
@@ -1,4 +1,4 @@
-package com.devonfw.application.quarkus.sample.animalmanagement.logic.api.to;
+package com.devonfw.demoquarkus.rest.v1.model;
 
 import lombok.Getter;
 import lombok.Setter;
