+++ reference-project/src/main/resources/application.properties
@@ -2,49 +2,56 @@
 # we can define profile-specific values using `%<profile>.` prefix
 # e.g. %dev.quarkus.log.level=DEBUG would only be applied in dev mode
 
-# drop and create the database at startup (other values: `validate`, `update`, `none`)
-# dont deploy this to any env you care about ;)
-quarkus.hibernate-orm.database.generation=drop-and-create
-# to print sql produced by hibernate, disable in prod
-quarkus.hibernate-orm.log.sql=true
+# Datasource configuration
+quarkus.datasource.db-kind=postgresql
+quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/demo
 
+# logging configuration
 # which packages should be automatically logged by tkit cdi logger
 quarkus.tkit.log.packages=org.tkit,com.devonfw
-
-#you can supress logs from classes by annotations or regex
+# you can supress logs from classes by annotations or regex
 quarkus.tkit.log.ignore.pattern=.*MapperImpl
 
-
-#identifier of your service, if empty a no-op tracer will be used
+# Tracing setup
+# identifier of your service, if empty a no-op tracer will be used
 quarkus.jaeger.service-name=demo-app
-#sampling algorithm, const=tracing all or nothing
+# sampling algorithm, const=tracing all or nothing
 quarkus.jaeger.sampler-type=const
-#1=trace every single method
+# 1=trace every single method
 quarkus.jaeger.sampler-param=1
-quarkus.log.console.format=%d{HH:mm:ss} %-5p traceId=%X{traceId}, parentId=%X{parentId}, spanId=%X{spanId}, sampled=%X{sampled} [%c{2.}] (%t) %s%e%n
-#report spans to central aggregator(you must start Jaeger yourself - e.g. using the docker-compose.yaml)
+# report spans to central aggregator(you must start Jaeger yourself - e.g. using the docker-compose.yaml)
 quarkus.jaeger.endpoint=http://localhost:14268/api/traces
 
-#all of these are default values, metrics would work just fine if you remove it
+# all of these are default values, metrics would work just fine if you remove it
 quarkus.micrometer.binder-enabled-default=true
-#also report metrics from microprofile
+# also report metrics from microprofile
 quarkus.micrometer.binder.mp-metrics.enabled=true
-#implicit metrics from http client
+# implicit metrics from http client
 quarkus.micrometer.binder.http-client.enabled=true
-#implicit metrics from REST endpoints
+# implicit metrics from REST endpoints
 quarkus.micrometer.binder.http-server.enabled=true
 
 
 # DEV profile
+# in dev mode, drop and create the database at startup (other values: `validate`, `update`, `none`)
+# dont ever enable this to any env you care about ;)
+%dev.quarkus.hibernate-orm.database.generation=drop-and-create
+# during dev, print sql produced by hibernate, disable in prod(default)
+%dev.quarkus.hibernate-orm.log.sql=true
 # disable json logs for better readability
 %dev.quarkus.tkit.log.console.json=false
 #if you dont use tkit-json=log but standard quarkus-json-log, use this instead
 #%dev.quarkus.log.console.json=false
+# if you want to see opentracing values in
+%dev.quarkus.log.console.format=%d{HH:mm:ss} %-5p traceId=%X{traceId}, parentId=%X{parentId}, spanId=%X{spanId}, sampled=%X{sampled} [%c{2.}] (%t) %s%e%n
 
 
 # TEST
-%test.quarkus.tkit.log.console.json=false
 
-quarkus.datasource.db-kind=postgresql
-quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/demo
-quarkus.datasource.credentials-provider=custom
+%test.quarkus.tkit.log.console.json=false
+# Test DB is provided by tkit test containers, URL will be set dynamically
+%test.quarkus.datasource.password=demo
+%test.quarkus.datasource.username=demo
+%test.quarkus.hibernate-orm.database.generation=drop-and-create
+%test.quarkus.hibernate-orm.log.sql=true
+
